requirements.txt
Eel==0.16.0
googletrans==4.0.0rc1
hugchat==0.4.8
playsound==1.2.2
pvporcupine==1.9.5
PyAudio==0.2.14
PyAutoGUI==0.9.54
pyttsx3==2.90
pywhatkit==5.4
SpeechRecognition==3.10.4
speedtest_cli==2.1.3
